[update] 즐겨찾기 기능 에러 수정
